requirements.txt
streamlit==1.41.1
python-docx==1.1.2
PyPDF2==3.0.1
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
plotly==5.24.1
openai==1.58.1
python-dotenv==1.0.1
